Address second-round strict-review findings
Thread padding_mask into the attention-only CUDA-graph replay: the graphed
overlap tail runs the MoE routing itself, but the schedule never put the mask
in kwargs, so it routed as if every batch were unpadded while eager did not.
Scope the threading to that path -- _replay_mhc_attention_consumer filters
kwargs through a tensor allowlist before the captured call, whereas the
generic replay hands **kwargs to TE, where an argument absent at capture time
would break replay. Match the eager callable exactly by dropping the
input_ids/packed_seq_params arguments it does not pass either.

Accumulate grad_h in fp32 in the torch h-aggregate backward so all three
implementations agree; the fallback is reachable whenever cuTile is absent,
including Triton-present builds.

Raise instead of assert in sync_before_step so the guard survives python -O,
where it is the only enforcement of the capture-scope invariant.

Reject full-iteration capture on the HybridStack mHC path: the config-level
gate that admits it is not model-family aware, and only the GPT path has been
validated.

Drop the unread ordinal and storage_offset bookkeeping from the arena slot
(data_ptr equality already subsumes offset equality), document why
BEFORE_MLP_BWD has no producer yet, list mhc_post in free_input_nodes
explicitly, move the slot-release NOTE above the loop it describes, and name
full-iteration capture in the overlap gate's error message.

+++ megatron/core/fusions/fused_mhc_kernels.py
@@ -2734,7 +2734,12 @@ def _get_triton_h_post_bda_bwd():
 def _torch_h_aggregate_bwd(grad_output: Tensor, x: Tensor, h_pre: Tensor) -> Tuple[Tensor, Tensor]:
     grad_output_expanded = grad_output.unsqueeze(2)
     grad_x = grad_output_expanded * h_pre.unsqueeze(-1)
-    grad_h = torch.sum(grad_output_expanded * x, dim=-1)
+    # Accumulate grad_h in fp32: the reduction spans the whole hidden dimension,
+    # and a bf16 product makes it ~3x noisier. The cuTile kernel
+    # (_ct_h_agg_bwd_kernel) and NativeHAggregateInto.backward both upcast for
+    # this reason, and this fallback is reachable whenever cuTile is absent --
+    # including Triton-present builds -- so it has to agree with them.
+    grad_h = torch.sum(grad_output_expanded.float() * x.float(), dim=-1)
     return grad_x.to(dtype=x.dtype), grad_h.to(dtype=h_pre.dtype)
